refactor: settings page
